ai-pc-agent/requirements.txt
# Core Dependencies
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0

# System Monitoring
psutil>=5.9.0
platformdirs>=3.0.0

# Browser Automation
playwright>=1.40.0

# Web Interface
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0

# File Management
watchdog>=3.0.0
python-magic>=0.4.27

# Notifications
discord.py>=2.3.0
python-telegram-bot>=20.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0

# Utilities
click>=8.1.0
tqdm>=4.65.0
rich>=13.0.0

# Optional: AI Integration
# openai>=1.0.0
# anthropic>=0.8.0
